Start implementing post-parser
Right now the type checker does a lot of parsing/grammar checking in
terms of matching the S-expressions into the actual grammar of the
language (converting exprs into vecs and strings, checking for
missing/extra arguments, etc.). What I expect is if I separate this
into two subtasks (parsing the S-expression tree into the
language-specific expression tree, and then type-checking), then the
type checker will be much easier to understand and work on,
as it will just be doing a single task. In addition, this refactoring
may allow more work to be reused for the main compiler.